Split up LQS by ledgers
