+++ application/ledger_query_service/src/bitcoin/bitcoind_zmq_listener.rs
