+++ application/ledger_query_service/src/bitcoin/queries.rs
