+++ application/ledger_query_service/src/ethereum/mod.rs
@@ -1,2 +1,2 @@
-pub mod bitcoind_zmq_listener;
 pub mod ethereum_web3_block_poller;
+pub mod queries;
