+++ application/ledger_query_service/src/ethereum/queries.rs
