--- application/ledger_query_service/src/queries/mod.rs
@@ -1,2 +1,0 @@
-pub mod bitcoin;
-pub mod ethereum;
